Adds a commondb package for shared db actions between plugins
