+++ src/plugins/jira-pond/package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/merico-dev/jira-pond#readme",
   "dependencies": {
-    "axios": "^0.21.1"
+    "axios": "^0.21.1",
+    "commondb": "file:../commondb"
   }
 }
